Refactor GildedRose
Extract an ItemDecorator class.
Create subclasses of the ItemDecorator for each item used in GildedRose.
Create an ItemFactory that creates the appropriate ItemDecorator based
on the name.
Replace the complex logic in update_quality with a call to tick for
each ItemDecorator instance.

_Massive_ commit - couldn't split this into hunks because of trailing
whitespace? I guess the original file was written on Windows machine...